Remove exception.
