+++ src/SerilogTestCorrelation/TestCorrelatorSink.cs
@@ -1,0 +1,13 @@
+﻿using Serilog.Core;
+using Serilog.Events;
+
+namespace SerilogTestCorrelation
+{
+    public class TestCorrelatorSink : ILogEventSink
+    {
+        public void Emit(LogEvent logEvent)
+        {
+            TestCorrelator.AddLogEvent(logEvent);
+        }
+    }
+}
